+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
   // New
   import { gl } from "npm:date-fns/locale/gl";
   ```
+- Allow to get value from attribute in CSS query of `getDataValue()` function. See `metas` plugin for a clear example.
 
 ### Fixed
 - Pages filtered with `filter_pages` plugin are exported to the sitemap.